Revert "refactor: SpringMotionProtocol for NSWindow and NSPanel conformance"
This reverts commit f552afb564f413558b0c28455a97f02720e06116.
